backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.7.0
python-multipart==0.0.17
httpx==0.27.2

SQLAlchemy==2.0.36
asyncpg==0.30.0
greenlet==3.1.1

pgvector==0.3.6

openai==1.58.1

python-dateutil==2.9.0.post0

pypdf==5.1.0
beautifulsoup4==4.12.3
lxml==5.3.0

# Audio (we won't fully use yet, but keep parity)
pydub==0.25.1